third_party/blink/web_tests/fast/forms/form-collection-elements-expected.txt

This test lists all the elements of a form per http://www.whatwg.org/specs/web-apps/current-work/multipage/forms.html#category-listed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

This test does not add object element as a listed element. Separate test will cover that.
PASS owner.elements.length is 22
PASS owner.elements[0].id is 'button1'
PASS owner.elements[1].id is 'fieldset1'
PASS owner.elements[2].id is 'inputhidden'
PASS owner.elements[3].id is 'inputtext'
PASS owner.elements[4].id is 'inputsearch'
PASS owner.elements[5].id is 'inputurl'
PASS owner.elements[6].id is 'inputemail'
PASS owner.elements[7].id is 'inputpassword'
PASS owner.elements[8].id is 'inputdate'
PASS owner.elements[9].id is 'inputnumber'
PASS owner.elements[10].id is 'inputrange'
PASS owner.elements[11].id is 'inputcolor'
PASS owner.elements[12].id is 'inputcheckbox'
PASS owner.elements[13].id is 'inputradio'
PASS owner.elements[14].id is 'inputfile'
PASS owner.elements[15].id is 'inputsubmit'
PASS owner.elements[16].id is 'inputreset'
PASS owner.elements[17].id is 'inputbutton'
PASS owner.elements[18].id is 'object1'
PASS owner.elements[19].id is 'output1'
PASS owner.elements[20].id is 'select1'
PASS owner.elements[21].id is 'textarea1'
PASS successfullyParsed is true

TEST COMPLETE

option1
